% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildParameters.R
\name{BuildParameters}
\alias{BuildParameters}
\title{BuildParameters}
\usage{
BuildParameters(par, model, data)
}
\arguments{
\item{par}{Vector of parameter values, transformed to have support on R.}

\item{model}{Natural history model of class \code{BCModel} as defined by \code{DefineBCModel}.}

\item{data}{Dataframe containing the covariate data of the model. One entry per individual.}
}
\value{

}
\description{
Predict BC risk with an upper limit on tumor size, both if and if not attending the next screening.
}
