% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcCensCase}
\alias{CalcCensCase}
\title{CalcCensCase}
\usage{
CalcCensCase(
  age,
  scr,
  par,
  t_pts,
  t_wts,
  r_pts,
  r_wts,
  d0 = 0.5,
  v0 = 0.06544985
)
}
\description{
Calculate the joint likelihood of a single censored
}
