% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcCensCaseLong}
\alias{CalcCensCaseLong}
\title{CalcCensCaseLong}
\usage{
CalcCensCaseLong(
  age,
  scr,
  par,
  scr_par,
  t_pts,
  t_wts,
  r_pts,
  r_wts,
  d0 = 0.5,
  v0 = 0.06544985
)
}
\description{
Calculate the likelihood of a single censored.
}
