% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcCensCaseODG}
\alias{CalcCensCaseODG}
\title{CalcCensCaseODG}
\usage{
CalcCensCaseODG(
  age,
  scr,
  par,
  t_pts,
  t_wts,
  r_pts,
  r_wts,
  d0 = 0.5,
  v0 = 0.06544985
)
}
\description{
Calculate the joint likelihood of a single censored, in a model with onset-dependent growth rates.
}
