% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcFutureScreeningSens}
\alias{CalcFutureScreeningSens}
\title{CalcFutureScreeningSens}
\usage{
CalcFutureScreeningSens(
  age,
  scr,
  par,
  t,
  t_wts,
  r,
  r_wts,
  d0 = 0.5,
  v0 = 0.06544985
)
}
\description{
Calculate the probability of being detected at a future screening. 
Note: the screening at \code{age} should not be included in the screening history \code{scr}.
}
