% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcIndividualLikelihood}
\alias{CalcIndividualLikelihood}
\title{CalcIndividualLikelihood}
\usage{
CalcIndividualLikelihood(
  is_case,
  is_scr,
  age,
  v,
  scr,
  entry,
  e_scr,
  par,
  t_pts,
  t_wts,
  r_pts,
  r_wts,
  d0 = 0.5,
  v0 = 0.06544985
)
}
\description{
Determine case and mode status, and calculate the corresponding joint likelihood.
}
