% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcIndividualLikelihoodODG}
\alias{CalcIndividualLikelihoodODG}
\title{CalcIndividualLikelihoodODG}
\usage{
CalcIndividualLikelihoodODG(
  is_case,
  is_scr,
  age,
  v,
  scr,
  entry,
  e_scr,
  par,
  t_pts,
  t_wts,
  r_pts,
  r_wts,
  d0 = 0.5,
  v0 = 0.06544985
)
}
\description{
Determine case and mode status, and calculate the corresponding joint likelihood, in a model with onset-dependent growth rates.
}
