% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcScreenCaseLong}
\alias{CalcScreenCaseLong}
\title{CalcScreenCaseLong}
\usage{
CalcScreenCaseLong(
  age,
  v,
  scr,
  par,
  scr_par,
  t_pts,
  t_wts,
  d0 = 0.5,
  v0 = 0.06544985
)
}
\description{
Calculate the joint likelihood of a single screen-detected case.
}
