% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcSymptCase}
\alias{CalcSymptCase}
\title{CalcSymptCase}
\usage{
CalcSymptCase(age, v, scr, par, t_pts, t_wts, d0 = 0.5, v0 = 0.06544985)
}
\description{
Calculate the joint likelihood of a single symptomatic case.
}
