% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcSymptCaseCumulative}
\alias{CalcSymptCaseCumulative}
\title{CalcSymptCaseCumulative}
\usage{
CalcSymptCaseCumulative(age, v, scr, par, t, t_wts, d0 = 0.5, v0 = 0.06544985)
}
\description{
Calculate the likelihood of being a symptomatic case, cumulative over age.
}
