% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcSymptCaseLong}
\alias{CalcSymptCaseLong}
\title{CalcSymptCaseLong}
\usage{
CalcSymptCaseLong(
  age,
  v,
  scr,
  par,
  scr_par,
  t_pts,
  t_wts,
  d0 = 0.5,
  v0 = 0.06544985
)
}
\description{
Calculate the joint likelihood of a single symptomatic case.
}
