% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineBCModel.R
\name{DefineBCModel}
\alias{DefineBCModel}
\title{DefineBCModel}
\usage{
DefineBCModel(
  onset = ~1,
  growth = ~1,
  sympt = ~1,
  sens = ~1,
  onset_dep_growth = FALSE,
  data = NULL,
  start_par = NULL
)
}
\arguments{
\item{onset}{Specifies covariates to include in the onset submodel.}

\item{growth}{Specifies covariates to include in the inverse growth rate submodel.}

\item{sympt}{Specifies covariates to include in the symptomatic detection submodel.}

\item{sens}{Specifies covariates to include in the screening sensitivity submodel.}

\item{onset_dep_growth}{Indicate if the inverse growth rate should depend on the age at onset (Strandberg et al. 2022).}

\item{data}{An optional dataframe representing the data the model is to be fitted to. Necessary if e.g. the formulas include spline functions. If provided, all predictor names in the four submodels must exist.}

\item{start_par}{Optionally specify (starting) parameter values.}
}
\value{
A list-type object of class \code{"BCmodel"}.
\describe{
  \item{$onset}{The specified formula for the onset submodel.}
  \item{$growth}{The specified formula for the growth rate submodel.}
  \item{$sympt}{The specified formula for the symptomatic detection submodel.}
  \item{$sens}{The specified formula for the screening sensitivity submodel.}
  \item{$onset_dep_growth}{Indicates if the mean of the inverse growth rate distribution depends on the age at onset}
  \item{$par}{Numeric vector containing parameter values for all four submodels, either provided by start.par or assigned by the function.}
  \item{$fitted}{Indicates if the model parameters have been fitted to data. Set to FALSE.}
}
}
\description{
Set up a natural history model.
}
\examples{
DefineBCModel()
DefineBCModel(sens = ~ percent_density, onset_dep_growth = TRUE)
DefineBCModel(onset = ~ family_history + birads, growth = ~ bmi)
}
