% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{EstimateBCModel}
\alias{EstimateBCModel}
\title{EstimateBCModel}
\usage{
EstimateBCModel(
  model,
  data,
  hessian = TRUE,
  n.cores = parallel::detectCores() - 1,
  d0 = 0.5,
  entry = "entry",
  exit = "exit",
  is.case = "case",
  mode = "mode",
  size = "size",
  scr.hist = "scr",
  gauss.kronrod.set = 6,
  gauss.leguerre.set = 4
)
}
\arguments{
\item{model}{Natural history model of class \code{BCModel} as defined by \code{DefineBCModel}.}

\item{data}{Data}

\item{hessian}{Indicates if the Hessian matrix should be returned.}

\item{n.cores}{Indicates the number of CPU threads to use.}

\item{d0}{The starting diameter of tumors at onset.}

\item{entry}{The variable name in \code{data} corresponding to the age at study entry.}

\item{exit}{The variable name in \code{data} corresponding to the age at end of follow-up.}

\item{is.case}{The variable name in \code{data} indicating BC case status (case = 1).}

\item{mode}{The variable name in \code{data} indicating the mode of detection (screen-detected = 1)}

\item{size}{The variable name in \code{data} corresponding to the tumor size (mm).}

\item{scr.hist}{The variable name in \code{data} corresponding to the screening histories (each entry a list of ages screened).}

\item{gauss.kronrod.set}{Description}

\item{gauss.leguerre.set}{Description}
}
\value{
Fitted Model
}
\description{
Fit a BCModel to cohort data.
}
