% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateBCModelLong.R
\name{EstimateBCModelLong}
\alias{EstimateBCModelLong}
\title{EstimateBCModelLong}
\usage{
EstimateBCModelLong(
  model,
  data,
  scr_data,
  link = NULL,
  e_link = NULL,
  age_mammogram = "age_mammogram",
  id = "id",
  base_variables = list(entry = "entry", exit = "exit", is_case = "case", mode = "mode",
    size = "size", scr_hist = "scr"),
  hessian = TRUE,
  n_cores = parallel::detectCores() - 1,
  d0 = 0.5
)
}
\arguments{
\item{model}{Natural history model of class \code{BCModel} as defined by \code{DefineBCModel}.}

\item{data}{Dataframe containing the cohort data. One entry per individual.}

\item{base_variables}{List linking variable names of \code{data} to variables used in the prediction: 
\describe{
  \item{entry}{The age at the start of the prediction.}
  \item{exit}{The age at end of prediction follow-up.}
  \item{is_case}{BC case status (case = 1).}
  \item{mode}{Mode of detection (screen-detected = 1).}
  \item{size}{tumor size (mm).}
  \item{scr_hist}{Screening histories (each entry a list of ages screened).}
}}

\item{hessian}{Indicates if the Hessian matrix of the maximum log-likelihood should be returned.}

\item{n_cores}{Indicates the number of CPU threads to use.}

\item{d0}{The starting diameter of tumors at onset.}

\item{gauss_kronrod_set}{Integer indicating how many nodes to use in Gauss-Kronrod quadrature. Values of 1-6 are supported and correspond to {11, 21, ..., 61} nodes, respectively.}

\item{gauss_leguerre_set}{Integer indicating how many nodes to use in Gauss-Leguerre quadrature. Values of 1-6 are supported and correspond to {10, 20, ..., 60} nodes, respectively.}
}
\value{
Fitted Model.
\describe{
  \item{$onset}{The specified formula for the onset submodel.}
  \item{$growth}{The specified formula for the growth rate submodel.}
  \item{$sympt}{The specified formula for the symptomatic detection submodel.}
  \item{$sens}{The specified formula for the screening sensitivity submodel.}
  \item{$onset_dep_growth}{Indicates if the mean of the inverse growth rate distribution depends on the age at onset}
  \item{$par}{Numeric vector containing parameter values for all four submodels, either provided by start.par or assigned by the function.}
  \item{$fitted}{Indicates if the model parameters have been fitted to data. Set to TRUE if the optimization succeeds.}
}
}
\description{
Fit a BCModel to cohort data.
}
