% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EvalCaseIntegral}
\alias{EvalCaseIntegral}
\title{EvalCaseIntegral}
\usage{
EvalCaseIntegral(age, v, scr, par, t, t_wts, d0 = 0.5, v0 = 0.06544985)
}
\description{
Evaluate the integral for BC cases.
}
