% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EvalCaseIntegralODG}
\alias{EvalCaseIntegralODG}
\title{EvalCaseIntegralODG}
\usage{
EvalCaseIntegralODG(age, v, scr, par, t, t_wts, d0 = 0.5, v0 = 0.06544985)
}
\description{
Evaluate the integral for BC cases, in a model with onset-dependent growth rates.
}
