% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EvalCensIntegralLong}
\alias{EvalCensIntegralLong}
\title{EvalCensIntegralLong}
\usage{
EvalCensIntegralLong(
  age,
  scr,
  par,
  scr_par,
  t,
  t_wts,
  r,
  r_wts,
  d0 = 0.5,
  v0 = 0.06544985
)
}
\description{
Evaluate the double integral for censored.
}
