% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EvalCensIntegralODG}
\alias{EvalCensIntegralODG}
\title{EvalCensIntegralODG}
\usage{
EvalCensIntegralODG(
  age,
  scr,
  par,
  t,
  t_wts,
  r,
  r_wts,
  d0 = 0.5,
  v0 = 0.06544985
)
}
\description{
Evaluate the double integral for censored, in a model with onset-dependent growth rates.
}
