% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchDataVariables.R
\name{MatchDataVariables}
\alias{MatchDataVariables}
\title{MatchDataVariabes}
\usage{
MatchDataVariables(
  data,
  base_variables = list(entry = "entry", exit = "exit", is_case = "case", mode = "mode",
    size = "size", scr_hist = "scr"),
  include_prevalent_screen = TRUE
)
}
\arguments{
\item{data}{Dataframe containing the data. One entry per individual.}

\item{base_variables}{List linking variable names of \code{data} to variables used in the prediction: 
\describe{
  \item{entry}{The age at the start of the prediction.}
  \item{exit}{The age at end of prediction follow-up.}
  \item{is_case}{BC case status (case = 1).}
  \item{mode}{Mode of detection (screen-detected = 1).}
  \item{size}{tumor size (mm).}
  \item{scr_hist}{Screening histories (each entry a list of ages screened).}
}}

\item{include_prevalent_screen}{Indicates if a screening performed at the time of entry should be included in the follow-up.}
}
\value{
\code{data.frame} with the following variables matched from the input data:
\describe{
  \item{$entry}{The age at the start of the prediction.}
  \item{$exit}{The age at end of prediction follow-up.}
  \item{$case}{BC case status (case = 1).}
  \item{$mode}{Mode of detection (screen-detected = 1).}
  \item{$v}{tumor volume (mm3).}
  \item{$scr}{Screening histories (each entry a list of ages screened).}
  \item{$e_scr}{Screening histories before study entry (each entry a list of ages screened).} }
}
\description{
Match a data set to the core variables used for estimation and prediction.
}
