% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictRegularRisk.R
\name{PredictRegularRisk}
\alias{PredictRegularRisk}
\title{PredictRegularRisk}
\usage{
PredictRegularRisk(
  model,
  data,
  base_variables = list(entry = "entry", exit = "exit", is_case = "case", mode = "mode",
    size = "size", scr_hist = "scr"),
  gauss_kronrod_set = 6,
  gauss_leguerre_set = 4,
  d0 = 0.5,
  n_cores = parallel::detectCores() - 1
)
}
\arguments{
\item{model}{Natural history model of class \code{BCModel} as defined by \code{DefineBCModel}.}

\item{data}{Dataframe containing the prediction data. One entry per individual.}

\item{base_variables}{List linking variable names of \code{data} to variables used in the prediction: 
\describe{
  \item{entry}{The age at the start of the prediction.}
  \item{exit}{The age at end of prediction follow-up.}
  \item{is_case}{BC case status (case = 1).}
  \item{mode}{Mode of detection (screen-detected = 1).}
  \item{size}{tumor size (mm).}
  \item{scr_hist}{Screening histories (each entry a list of ages screened).}
}}

\item{gauss_kronrod_set}{Integer indicating how many nodes to use in Gauss-Kronrod quadrature. Values of 1-6 are supported and correspond to {11, 21, ..., 61} nodes, respectively.}

\item{gauss_leguerre_set}{Integer indicating how many nodes to use in Gauss-Leguerre quadrature. Values of 1-6 are supported and correspond to {10, 20, ..., 60} nodes, respectively..}

\item{d0}{The starting diameter of tumors at onset.}

\item{n_cores}{Indicates the number of CPU threads to use.}
}
\value{
\code{data.frame} with predictions of breast risk.
\describe{
  \item{$age}{Age at the start of the prediction interval.}
  \item{$end}{Age at the end of the prediction interval.}
  \item{$surv}{The model fitted probability of surviving until \code{age}.}
  \item{$risk_attend}{The predicted risk of diagnosis, if attending the next screening.}
  \item{$risk_noattend}{The predicted risk of diagnosis, if not attending the next screening.}
}
}
\description{
Predict BC risk, both if and if not attending the next screening.
}
