% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{SummarizeBCEstimates}
\alias{SummarizeBCEstimates}
\title{SummarizeBCEstimates}
\usage{
SummarizeBCEstimates(model, alpha = 0.05, trans = TRUE)
}
\arguments{
\item{model}{Natural history model of class \code{BCModel} as defined by \code{DefineBCModel}.}

\item{alpha}{Par.}

\item{trans}{Par.}
}
\value{
Fitted Model.
}
\description{
Description
}
