% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{summary.BCModel}
\alias{summary.BCModel}
\title{summary.BCModel}
\usage{
\method{summary}{BCModel}(model, ...)
}
\arguments{
\item{model}{Natural history model of class \code{BCModel} as defined by \code{DefineBCModel}.}
}
\value{
Fitted Model.
}
\description{
Description
}
