% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{EstimateBCModel}
\alias{EstimateBCModel}
\title{EstimateBCModel}
\usage{
EstimateBCModel(
  model,
  data,
  hessian = TRUE,
  n_cores = parallel::detectCores() - 1,
  d0 = 0.5,
  entry = "entry",
  exit = "exit",
  is_case = "case",
  mode = "mode",
  size = "size",
  scr_hist = "scr",
  gauss_kronrod_set = 6,
  gauss_leguerre_set = 4
)
}
\arguments{
\item{model}{Natural history model of class \code{BCModel} as defined by \code{DefineBCModel}.}

\item{data}{Dataframe containing the cohort data. One entry per individual.}

\item{hessian}{Indicates if the Hessian matrix should be returned.}

\item{n_cores}{Indicates the number of CPU threads to use.}

\item{d0}{The starting diameter of tumors at onset.}

\item{entry}{The variable name in \code{data} corresponding to the age at study entry.}

\item{exit}{The variable name in \code{data} corresponding to the age at end of follow-up. For BC cases, this is also the age at diagnosis.}

\item{is_case}{The variable name in \code{data} indicating BC case status (case = 1).}

\item{mode}{The variable name in \code{data} indicating the mode of detection (screen-detected = 1).}

\item{size}{The variable name in \code{data} corresponding to the tumor size (mm).}

\item{scr_hist}{The variable name in \code{data} corresponding to the screening histories (each entry a list of ages screened).}

\item{gauss_kronrod_set}{Integer indicating how many nodes to use in Gauss-Kronrod quadrature. Values of 1-6 are supported and correspond to {11, 21, ..., 61} nodes, respectively.}

\item{gauss_leguerre_set}{Integer indicating how many nodes to use in Gauss-Leguerre quadrature. Values of 1-6 are supported and correspond to {10, 20, ..., 60} nodes, respectively..}
}
\value{
Fitted Model.
}
\description{
Fit a BCModel to cohort data.
}
